+++ src/main/java/com/educate/creationofdatabase/domain/model/TypesEmployees.java
@@ -1,23 +1,23 @@
 package com.educate.creationofdatabase.domain.model;
 
 public class TypesEmployees {
-    private Integer id;
+    private Long id;
     private String name;
 
     public TypesEmployees() {
 
     }
 
-    public TypesEmployees(Integer id, String name) {
+    public TypesEmployees(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
